Services fix

+++ src/main/java/com/example/Betting/service/IUserService.java
@@ -22,8 +22,8 @@ public interface IUserService {
      *
      * @param id the id
      * @param transactionMoney the transaction money
-     * @param adding the adding
+     * @param adding the adding money value
      * @return the optional user
      */
-    Optional<User> changeMoneyValueInWallet(Long id, float transactionMoney, boolean adding);
+    Optional<User> changeMoneyValueInWallet(Long id, int transactionMoney, boolean adding);
 }
